moved context export into hook
