+++ src/contexts/web3Data/hooks/useWeb3Provider.ts
@@ -1,5 +1,16 @@
-import { useContext } from 'react';
-import { IWeb3ProviderContext, Web3ProviderContext } from '../Web3ProviderContext';
+import { createContext, useContext } from 'react';
+
+import { ActionTypes } from '../actions';
+import { ConnectFn, DisconnectFn, InitialState } from '../types';
+
+export interface IWeb3ProviderContext {
+  state: InitialState;
+  dispatch: React.Dispatch<ActionTypes>;
+  connect: ConnectFn;
+  disconnect: DisconnectFn;
+}
+
+export const Web3ProviderContext = createContext<IWeb3ProviderContext | null>(null);
 
 export const useWeb3Provider = (): IWeb3ProviderContext =>
   useContext(Web3ProviderContext as React.Context<IWeb3ProviderContext>);
